Refactor: generateRandomPostFix 클래스화
